moved unit tests for getElementsByIndexLessThan() and getElementsByIndexGreaterThan() to ElementTest.java
